services/ga/requirements.txt
analytics-mcp==0.2.0
mcp>=1.24.0
authlib>=1.6.0
httpx>=0.28.1
starlette>=0.52.0
uvicorn>=0.42.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
